Move tendermint's types to types file
